moved pet to its own component, installed React, added .gitignore
