+++ .gitignore
@@ -1,0 +1,7 @@
+node_modules/
+.cache
+dist/
+.env
+.DS_Store
+coverage/
+.vscode/
