+++ src/index.html
@@ -10,9 +10,6 @@
   <body>
     <!-- React's entry point -->
     <div id="root">Not Rendered</div>
-    <!-- below is the React package we'll use. React uses react-dom for rendering the virtual dom -->
-    <script src="https://unpkg.com/react@17.0.1/umd/react.development.js"></script>
-    <script src="https://unpkg.com/react-dom@17.0.1/umd/react-dom.development.js"></script>
     <script src="App.js"></script>
   </body>
 </html>
